fix: firebaseの設定を別のファイルに独立させるため、firebase.config.jsを作成
